stop key down event listener
There is a bug on the code where if the user press any key before the
countdown finishes the input gets registered and the first letter to be
displayed gets marked as right or wrong.

This commit fixes this bug.

When the game starts after the countdown the key down event listener is
added to capture the user's input.
Once the game finishes the event listener is removed.